Save data for only the most recent attempt of each exercise
Based on user feedback, there was no need to keep a record of each exercise attempt.  Users want to see which exercises have been completed and which have not.  A student's most recent code and submission date is saved for each attempted exercise.  This significantly reduces the number of records in the database.