Fix bug in User#show method computing last attempt
If a student had no attempted exercises or challenges, the User#show controller was defining @last_attempt as [nil].  The view was checking whether or not the array was empty.  Because it contained this nil value, the array was passed to the last_submisssion_date helper method, which attempted to call submitted at on the NilClass object, causing an error that prevented users from being able to view the pages of students with no challenge and exercise attempts.