Value for NYC cohort location updated
The helper method that divides cohorts by location for the sidebar was looking for the NYC cohorts to have the location "New York", but it is "New York City".  This difference was causing the app to
crash.

Also, sorted the cohorts so that they are listed at each location in order of start date.