Save cohort names with three words correctly
Cohort names are entered into Socrates as name-space-year (e.g., Squirrels 2014).  Before cohorts are saved in the database, the ordering is altered to year-space-name (e.g. 2014 Squirrels) so that they display well in the sidebar.

The regex used to make the alteration only handled cohort names with one or two words plus the year (e.g., Banana Slugs 2014).  It did not work with the three-word NYC cohorts (e.g., Woolly Bear Caterpillars 2014).  Rather than being saved in the database as "2014 Woolly Bear Caterpillars", the name was saved as "Woolly 2014 Bear Caterpillars".